Cleaned up SolitaireGame and CardManager classes by moving all board related logic to the board class
